+++ src/mkvtoolnix-gui/header_editor/page_model.h
@@ -48,6 +48,10 @@ public:
 
 Q_SIGNALS:
   void attachmentsReordered();
+  void tracksReordered();
+
+public Q_SLOTS:
+  void rereadTopLevelPageIndexes();
 };
 
 }